Add a utility function for extracting exclusion tags
In order to be able to verify exclusion tag violations in cluster
verify, we need to decide which tags are exclusion tags. So add
two python utility functions: one for extracting exclusion prefixes
from from the cluster tags and one testing if a tag starts with
one of a set of given prefixes.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>