List already submitted jobs on user cancel interrupt
Currently only not yet submitted jobs are showed. This patch
Adds printing already submitted jobs, their ID's and statuses on
user cancel interrupt. That is requested in Issue 357.

Submitted jobs, their ID's and statuses are
accumulated in 'submitted' list. This way we only accumulate jobs
submitted by execWithCancel. The problem is that jobs submitted
directly via submitJobs are not accumulated. But that's not really
a problem because submitJobs is used only from Harep and in that
case even cancel signal handler is not registered because there is
no execWithCancel call in Harep.

Signed-off-by: Oleg Ponomarev <onponomarev@gmail.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>