Consider N+1 redundancy for non-DRB iterated allocation
As described in our design, we keep iterated allocation as is for
DRBD instances; there redundancy is a local property, affecting
only neighbor nodes. Also, we keep the semantics of "accept existing"
and "independent groups". For shared-storage instances, however, we
take N+1 redundancy into account as we do for normal allocation.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>