writeConfigAndUnlock: only unlock after config write
Otherwise the writing of the unlock can overtake the
configuration writing and if the wconfd gets killed
forcefully in that moment an incorrect state is left
on disk. If the calling job was not affected, then
once it will retry it will believe it succeeded already
because the lock is gone, hence leaving the configuration
in an inconsistent state.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>