QA: use relative path to script location for burnin
The QA had the annoying bug that if you run burnin, it
would try to upload the burnin file from the tools/
directory relative to the working directory and not
the location of the qa_cluster.py script. If you
accidentally start the QA from a differen directory
than qa/ it would run for a long while and only
crash when it comes to the burnin part. This patch
fixes the glitch by taking the location of the
QA scripts as a reference rather than the working
directory.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>