Avoid watcher interaction in TestFilterRateLimit
This test verifies a global limit on the number of jobs running. This
requires knowledge of all jobs submitted between the addition of the
filter and the submission of our last test job. While we send these
commands directly one after the other, this still takes a second or
two, thus giving the watcher slightly less than a 1% change of
interfering. Avoid this by pausing the watcher during this test.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>