Add lenses for simple fields
Lenses are currently not generated for
simple fields, as their types in the forthcoming
and real variants are not equal (And hence no simple
lens can be generated). However we can still generate
a more complex lens of type Lens s s (Maybe a) a.

This will let us use set and over (but now view) for
these fields.

Signed-off-by: BSRK Aditya <bsrk@google.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>