Fix allocator call with empty whitelist
Our calling convention for the iallocator is, that if a non-None
value for the node_whitelist is passed, only nodes from this list
should be allocated. While it is not useful to pass the empty list
here, the allocator should still dutifully fail instead of choosing
from the whole set of nodes.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>