Fix groups.in to actually contain groups
While we have never actually heard of anyone using it, Ganeti does
offer the possibility of using a different prefix for the groups and
users used in split-user mode. The groups.in file contained a mishmash
of users and groups, and this did not work well with different
prefixes. This patch fixes the situation.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>