Introduce socat as a way of doing xl migrations
This patch introduces support for socat as a means of doing xl
migrations. The primary reason for doing so is that Ganeti no longer
handles SSH key distribution across nodes which are not master
candidates. By relying on SSH as the only means of doing migrations,
we could only migrate instances off of master candidates.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>