During start, only update job file locally
The first thing the newly forked job will do anyway is
to update its job file as the first op-code is now waiting.
It is enough that this information is replicated to all
master candidates. Note that a local change of the livelock
file will not be helpful to other master candidates, as
the livelock has only a meaning on the current node anyway.

This seemingly little saving of two replications is significant,
however, as it happens under the fork lock by which we avoid
two forks at the same time, as this can cause problems with the
Haskell runtime.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>

Cherry-picked-from: 7684a50192bb
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>