Export updateStatusAndScheduleSomeJobs
This function, when called, will do the same tasks that are
done periodically in the job queue anyway: the status of all
jobs considered running is verified and some jobs are scheduled,
should this be possible.

This will allow to react quicker to events indicating that
a burst of untracked jobs has finished.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>