Attempt to cleanup failed migrations using a pidfile
In the case that a listening socat daemon was started but the migration
failed on the sending side, the daemon will stay in place and occupy
the migration port forever. This patch attempts to remedy this by
saving the PID of the daemon, and attempting to kill it when the next
migration is started, provided the command line roughly matches our
migration workflow.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>