Add a function to find a suitable secondary
On disk conversion to DRBD, a new secondary node has to be found.
Add a function doing this search in a way similar to what instance
allocation does: all possible placements are considered and the
one resulting in the best cluster metrics is taken.

To have a uniform interface, use the same signature as all allocation
functions, even though the group list is not needed (we have to take
the group of the primary node anyway).

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>