Improve the speed of iterateAlloc
In the current computation of `iterateAlloc` most of the time
is spent in checking whether the obtained situation is globally
N+1 redundant. However, actually finding a non N+1 redundant
situation only happens at the very end in a long sequence of iterations
in an hspace-like use case. Therefore, we allocate a group of instances
in one block without checking global N+1 redundancy; if that happens to
yield a globally N+1 redundant situation, we can continue from there.
Only when that heuristics fails, we do allocation step by step checking
global N+1 redundancy in each allocation step.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>