Add option --no-capacity-checks to hspace
While it is generally a good idea to plan capacity to have a redundant
cluster, there are legitimate use cases, where we want to know how many
instances can be squeezed on the cluster at all, taking only N+1 redundancy
for DRBD into account. So add an option restoring the old behavior.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>