Add a new deamon for maintenance
This daemon will take over cluster maintenance as per
our design document. As it will heavily depend on the monitoring
daemon, it will only be enabled (at configure time) if the
monitoring daemon is enabled as well. It will also run as
the same user and group.

In this commit only the plain daemon is added with the only supported
request being the question for the supported protocol versions.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>