Add test showing a bug in location score calculation
The bug is using incorrect location score during failover and replace
action. Thus the idea of the test is to introduce a situation in which
it's profitably to perform failover and replace move in order to
optimize cluster score (if ignoring location tags). Considering location
tags another final placement is optimal.

In this test the optimal placement should locate primary instances to
nodes 1 and 3 with secondaries on the second node.

Signed-off-by: Oleg Ponomarev <onponomarev@gmail.com>
Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>