Improve speed of iterateAlloc by a trying two step sizes
Currently, iterateAlloc tries one guess on the remaining capacity
and falls back to small steps if that guess turns out to be too
optimistic. In the typical case, that the allocation is bound by
memory that initial guess works quite well; however, in some cases
other requirements limit the amount of instances allocatable on a
cluster. Instead of immediately giving up in this case, try smaller
guess-and-verify steps to avoid having to check for global N+1 redundancy
too often.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>