Extend maintenance configuration by balancing
One of the tasks the maintenance daemon does is keeping
the cluster balanced, taking load into account. As this
is not wanted for all use cases of Ganeti, there is a
flag to control it. Also, to avoid flapping behavior,
a threshold for balancing operations can be specified;
a balancing move will only be carried out if the gain
in the cluster metrics is bigger than the given threshold.
We also provide safe default values: balancing is
disabled by default, and the threshold is 10 times higher
than the one for manual balancing.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>