Redistribute master key on downgrade
When downgrading from 2.13 to 2.12, the SSH key pair of
each node needs to be replaced by the master's SSH key
pair. If that is not done, any node added after the
downgrade will not be reachable if the master is failed
over to one of the original non-master nodes.

This patch fixes Issue 1008. However, what is not cleaned
up is the nodes' authorized_keys file. This will need
significantly more effort, but this patch at least fixes
the user-impacting issue.

Signed-off-by: Helga Velroyen <helgav@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>