Support sphinx 1.3
First, enable_manpages is now required to be a bool; fortunately,
we set it via the environment in conf.py anyway, so no need to
pass it as an option as well. Also, the default template has been
renamed to classic; so branch on the sphinx version to choose the
correct name. Fixes issue #1119.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>