Fix rollover condition in xenload
The Xen CPU data collector gets as raw data accumulated CPU usage
since an arbitrary start date. If the value gets too large, the
counter rolls over. The XenCpuLoad collector is aware of this
happening and adjusts the observed values accordingly. However,
the condition for detecting a rollover was wrong: on an rollover,
the value strictly goes down. The value can stay the same over
an interval, e.g., for very idle guests; not that the accuracy
of the reported accumulated value can be as coarse as a whole
CPU second.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>