In multi-group allocation think about the whole cluster
When choosing a group for allocating a new instance, not only
consider how balanced the group of the new instance would be,
but look at the overall balancedness of the cluster.

The net effect of this change is that hail will prefer the most
empty group and not (as it is now) the most balanced node group.
This will keep the overall cluster more balanced.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>