Make the Xen CPU collector keep old data longer
The Xen CPU collector keeps old values of instance loads for some
time, even after they have moved away. In this way, we can still
obtain a buest-guess load values for an instance before a new load
value can be established on the new node.

For "replace primary" moves on an instance, however, the instance will
visit 3 nodes in short succession. Therefore, keeping old data only
for the time it takes to establish a new load on one node is not
necessarily enough; the instance could be on the intermediate node for
too short to establish a stable load value there. The theoretical
minimum would be 2 times the threshold for accepting load data as
stable; however, due to differences in the sampling time, play it safe
and use 3 times this value. Note that balancing will prefer the load
data from the node where the instance resides, whenever available; so
no harm is done to the accuracy by keeping old values a bit longer.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>