Speed up balancing by accepting pre-migration loads
It is perfectly sufficient to wait for new load data to
be collected on the new node for a just moved instance before
continuing with any further balancing operations (the data
will stabilize eventually). However, for quick moves, e.g.,
when evacuating a drained move, such delay can be annoying.
Therefore, also accept pre-migration load data for the instance
which can be found on other nodes; this is justified, as the
monitoring daemons will clean up old load data in a timely
fashion and most likely an instance will continue what it
was doing after a move.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>