Extend maintenance configuration by evacuated instances
When collecting instance load on a per-instance basis (currently Xen only),
There is one case where we do not have load information at all for an
instance: an instance just evacuated from an offline node. On the other
hand, this is precisely the situation where we do not want to wait for
load data to come in---an offline node should be evacuated as soon as
possible. Therefore, we need keep track of the instances we recently
evacuated so that we can accept missing load data for them.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>