tieredAlloc: in case of FailN1, shrink memory
When computing tiered allocation statistics, the normal
step is to shrink the resource showing the most errors.
However, for some abstract resources, like N+1 redundancy,
there is no physical resource this concept refers to;
nevertheless, there is an underlying physical resource that
most likely causes this kind of failure. For N+1 redundancy,
the missing resource almost always is memory. So shrink
based on this assumption.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>