tieredAlloc: Shrink for effect
When shrinking an instance do not shrink just in arbitrary
"unit" steps, but instead shrink until the failure statistics
changes. In this way, we only consider side branches (unilaterally
changing a single resource despite not the reason for the
majority of failures) only at places where we there is the possibility
to achieve something.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>