Relax test to only test for specified properties
The 7th shell test for hspace relied on unspecified
behavior: tiered allocatio requires the resource causing
the largest number of allocations to be immpossible to
be shrunk first, if that leads to a valid allocation.
However, in this example, the actual failure statistics
is
  [(FailMem,2),(FailDisk,0),(FailCPU,0),(FailN1,0),(FailTags,0)
  ,(FailMig,0),(FailDiskCount,0),(FailSpindles,2),(FailInternal,0)]
So it is up to the implementation to choose on whether to shrink
spindles or memory first. Therefore, accept the results of both
shrinking orders.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>