Register correct jobs for harep task in maintd
After submitting the jobs for harep-style repair in the
maintenance daemon, we register the jobs we submitted.
To avoid serial bumps, we only do this, if the change
is a no-op, i.e., if we didn't submit any jobs. To check
for this, we have to look at the new jobs, not the old
ones. Fix this.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>