Collect diagnose data from MonDs
At the beginning of each maintenance round, collect the
data for the self-diagnose collector from each node. If
the state is not Ok, and the incident is not known already,
register it. Note that we poll from each online node even
if an incident is known already, as the state of the node
might have gotten worse.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>