gnt-node add: password auth is only one method
The man page for gnt-node add is slightly misleading.
In can be read to mean that we insist on password
authentication for root on the target node. However,
the only point the man page is trying to make here
is that credentials have to be provided to allow to
ssh as root to the target node; password is only of
the many options. Fix the wording accordingly.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>