Only insist on haskell test dependencies if needed
While developers usually run tests and hence need all the infrastructure
for it, people who just want to build and use Ganeti usually do not want
to run tests. So, while still using cabal to get a consistent set of
dependency packages, make the test infra structure optional in the cabal
file. The decision on whether to require test dependencies will be made
based on whether the haskell tests are configured to be run.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>