Remove obvious debug message
When running Ganeti in debug mode, certain messages appear more than
others. The "up-to-date config" message is a very common sight and a
cause of clutter in cluster logs. This patch removes it because it is
not useful - the control flow contains meaningful messages for all
other cases (config request, WConfd polling), and the absence of any
message is equally indicative as the presence of this one.

Signed-off-by: Hrvoje Ribicic <riba@google.com>
Reviewed-by: Klaus Aehlig <aehlig@google.com>