Implement clean up of incidents
Once the handling of an incident is finished (either successfully or
failed), the corresponding node is tagged accordingly and the state
is set to success or failure. It is then the task of the user to confirm
that all needed non-Ganeti actions are taken by removing that tag.
Once the tag is removed, the repair daemon will forget the event.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Helga Velroyen <helgav@google.com>