Add a design for memory over-commitment
While Ganeti traditionally is very careful about making
sure no instance every runs out of memory. However, there
are situations where it can be desirable to accept parts of
the memory to be swapped out, e.g., if, from the use case,
it is known that always some machines will have memory not
actively accessed. Add a design for this situation that by
default preserves the current state of affairs.

Signed-off-by: Klaus Aehlig <aehlig@google.com>
Reviewed-by: Hrvoje Ribicic <riba@google.com>