Reintroduce FromString, and replace Error with it
We once had a class, FromString, that duplicated
the functionality of the libary given class Error.

We have since then replaced all uses of FromString
with Error. Now, Error is deprecated (since mtl >=
2.2.1). Hence, we are reintroducing FromString,
and are replacing all uses of Error with it.

The patch that removed FromString in favor
of Error is: a87a017b023e2979b73f11dcf9012602bfa4b13c

Signed-off-by: Bhimanavajjula Aditya <bsrk@google.com>
Signed-off-by: Petr Pudlak <pudlak@google.com>
Reviewed-by: Petr Pudlak <pudlak@google.com>